Team descriptions can be null
When a team does not have a description, GitHub's API sets the
`description` field to `null`. This breaks deserialization, which
expects the field to be set to a `String`.

This is most likely only an issue for teams that have not previously
been managed by sync-team, e.g. when an existing organizations is added
to the team repository.