Change repository description to not be optional
There was kind of a double empty state for repository descriptions (`None` or `Some("")`), which was causing unnecessary diffs to be displayed in logs (and performed in the API).

Missing or empty description is pretty much the same, and furthermore `team` requires the description field on each repository, so it shouldn't even ever be completely missing anyway (although it might be empty).